Moved Traits into directory and modify the users to point to them
