+++ app/Models/LicenseSeat.php
@@ -4,6 +4,7 @@ namespace App\Models;
 
 use App\Models\Traits\Acceptable;
 use App\Models\Traits\CompanyableChildTrait;
+use App\Models\Traits\Loggable;
 use App\Notifications\CheckinLicenseNotification;
 use App\Notifications\CheckoutLicenseNotification;
 use App\Presenters\Presentable;
