+++ app/Models/Traits/Requestable.php
@@ -1,8 +1,10 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Traits;
 
 use Illuminate\Support\Facades\Auth;
+use App\Models\CheckoutRequest;
+use App\Models\User;
 
 // $asset->requests
 // $asset->isRequestedBy($user)
